requirements.txt
Django==4.2.11 # Пингуем до последней стабильной версии 4.2.x
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.3
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.7
dj-database-url>=2.1.0
python-decouple>=3.8
drf-spectacular>=0.26.5
django-modeltranslation>=0.18.11
django-extensions>=3.2.3
Pillow>=10.0.0
requests>=2.31.0 # Оставляем пин из рабочего проекта
python-telegram-bot>=20.0
resend>=0.6.0
django-anymail>=10.2
django-redis>=5.4.0
redis>=5.0.0
celery>=5.3.0
python-slugify>=8.0.1
gunicorn>=21.2.0
whitenoise>=6.6.0
